Refactor the package structure
