+++ app/src/main/java/com/pragati/karuna/home/ui/GalleryFragment.kt
@@ -1,4 +1,4 @@
-package com.pragati.karuna.ui.gallery
+package com.pragati.karuna.home.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,7 +9,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.pragati.karuna.R
-import com.pragati.karuna.viewmodel.GalleryViewModel
+import com.pragati.karuna.home.viewmodel.GalleryViewModel
 
 class GalleryFragment : Fragment() {
