+++ app/src/main/java/com/pragati/karuna/home/viewmodel/GalleryViewModel.kt
@@ -1,13 +1,13 @@
-package com.pragati.karuna.viewmodel
+package com.pragati.karuna.home.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class GalleryViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
