+++ app/src/main/java/com/pragati/karuna/request/repository/RequestRepository.kt
@@ -1,9 +1,9 @@
-package com.pragati.karuna.repository
+package com.pragati.karuna.request.repository
 
 import android.util.Log
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
-import com.pragati.karuna.models.Request
+import com.pragati.karuna.request.model.Request
 
 class RequestRepository {
     private val db = Firebase.firestore
