+++ app/src/main/java/com/pragati/karuna/request/viewmodel/LocationViewModel.kt
@@ -1,8 +1,8 @@
-package com.pragati.karuna.viewmodel
+package com.pragati.karuna.request.viewmodel
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.pragati.karuna.models.Location
+import com.pragati.karuna.request.model.Location
 
 class LocationViewModel : ViewModel() {
     val location = MutableLiveData<Location>()
