+++ app/src/main/res/layout/fragment_add_kit.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:layout_marginStart="16dp"
     android:layout_marginEnd="16dp"
-    tools:context=".ui.kit.AddKitFragment">
+    tools:context=".request.ui.AddKitFragment">
 
     <Spinner
         android:id="@+id/sp_kit_type"
